Move table test helper function to component file
